@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.27.1
-// 	protoc        v3.6.1
+// 	protoc-gen-go v1.36.7
+// 	protoc        v3.21.12
 // source: twt.proto
 
 package twtproto
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -25,10 +26,12 @@ type ProxyComm_MessageType int32
 const (
 	ProxyComm_DATA_DOWN    ProxyComm_MessageType = 0 // downlink data, presumably most frequent message type
 	ProxyComm_DATA_UP      ProxyComm_MessageType = 1 // uplink data
-	ProxyComm_OPEN_CONN    ProxyComm_MessageType = 2 // open remote connection to server
-	ProxyComm_CLOSE_CONN_S ProxyComm_MessageType = 3 // close remote connection to server
-	ProxyComm_CLOSE_CONN_C ProxyComm_MessageType = 4 // close local connection to client
-	ProxyComm_PING         ProxyComm_MessageType = 5 // ping
+	ProxyComm_ACK_DOWN     ProxyComm_MessageType = 2 // acknowledge reception of downlink data
+	ProxyComm_ACK_UP       ProxyComm_MessageType = 3 // acknowledge reception of uplink data
+	ProxyComm_OPEN_CONN    ProxyComm_MessageType = 4 // open remote connection to server
+	ProxyComm_CLOSE_CONN_S ProxyComm_MessageType = 5 // close remote connection to server
+	ProxyComm_CLOSE_CONN_C ProxyComm_MessageType = 6 // close local connection to client
+	ProxyComm_PING         ProxyComm_MessageType = 7 // ping
 )
 
 // Enum value maps for ProxyComm_MessageType.
@@ -36,18 +39,22 @@ var (
 	ProxyComm_MessageType_name = map[int32]string{
 		0: "DATA_DOWN",
 		1: "DATA_UP",
-		2: "OPEN_CONN",
-		3: "CLOSE_CONN_S",
-		4: "CLOSE_CONN_C",
-		5: "PING",
+		2: "ACK_DOWN",
+		3: "ACK_UP",
+		4: "OPEN_CONN",
+		5: "CLOSE_CONN_S",
+		6: "CLOSE_CONN_C",
+		7: "PING",
 	}
 	ProxyComm_MessageType_value = map[string]int32{
 		"DATA_DOWN":    0,
 		"DATA_UP":      1,
-		"OPEN_CONN":    2,
-		"CLOSE_CONN_S": 3,
-		"CLOSE_CONN_C": 4,
-		"PING":         5,
+		"ACK_DOWN":     2,
+		"ACK_UP":       3,
+		"OPEN_CONN":    4,
+		"CLOSE_CONN_S": 5,
+		"CLOSE_CONN_C": 6,
+		"PING":         7,
 	}
 )
 
@@ -79,26 +86,23 @@ func (ProxyComm_MessageType) EnumDescriptor() ([]byte, []int) {
 }
 
 type ProxyComm struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Mt            ProxyComm_MessageType  `protobuf:"varint,1,opt,name=mt,proto3,enum=twtproto.ProxyComm_MessageType" json:"mt,omitempty"`
+	Proxy         uint64                 `protobuf:"varint,2,opt,name=proxy,proto3" json:"proxy,omitempty"`           // id of the proxy which is sending the message, must by globally unique
+	Connection    uint64                 `protobuf:"varint,3,opt,name=connection,proto3" json:"connection,omitempty"` // sequential number of the connection on the local side of the proxy
+	Seq           uint64                 `protobuf:"varint,4,opt,name=seq,proto3" json:"seq,omitempty"`               // sequential number of the message within the connection, sequential numbers are independent per up/down directions
+	Address       string                 `protobuf:"bytes,5,opt,name=address,proto3" json:"address,omitempty"`        // remote address to connect to, only used in OPEN_CONN message
+	Port          uint32                 `protobuf:"varint,6,opt,name=port,proto3" json:"port,omitempty"`             // remote port to connect to, only used in OPEN_CONN message
+	Data          []byte                 `protobuf:"bytes,7,opt,name=data,proto3" json:"data,omitempty"`              // data chunk, only used in DATA_* messages
 	unknownFields protoimpl.UnknownFields
-
-	Mt         ProxyComm_MessageType `protobuf:"varint,1,opt,name=mt,proto3,enum=twtproto.ProxyComm_MessageType" json:"mt,omitempty"`
-	Proxy      uint64                `protobuf:"varint,2,opt,name=proxy,proto3" json:"proxy,omitempty"`           // id of the proxy which is sending the message, must by globally unique
-	Connection uint64                `protobuf:"varint,3,opt,name=connection,proto3" json:"connection,omitempty"` // sequential number of the connection on the local side of the proxy
-	Seq        uint64                `protobuf:"varint,4,opt,name=seq,proto3" json:"seq,omitempty"`               // sequential number of the message within the connection, sequential numbers are independent per up/down directions
-	Address    string                `protobuf:"bytes,5,opt,name=address,proto3" json:"address,omitempty"`        // remote address to connect to, only used in OPEN_CONN message
-	Port       uint32                `protobuf:"varint,6,opt,name=port,proto3" json:"port,omitempty"`             // remote port to connect to, only used in OPEN_CONN message
-	Data       []byte                `protobuf:"bytes,7,opt,name=data,proto3" json:"data,omitempty"`              // data chunk, only used in DATA_* messages
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ProxyComm) Reset() {
 	*x = ProxyComm{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_twt_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_twt_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ProxyComm) String() string {
@@ -109,7 +113,7 @@ func (*ProxyComm) ProtoMessage() {}
 
 func (x *ProxyComm) ProtoReflect() protoreflect.Message {
 	mi := &file_twt_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -175,47 +179,45 @@ func (x *ProxyComm) GetData() []byte {
 
 var File_twt_proto protoreflect.FileDescriptor
 
-var file_twt_proto_rawDesc = []byte{
-	0x0a, 0x09, 0x74, 0x77, 0x74, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x08, 0x74, 0x77, 0x74,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0xae, 0x02, 0x0a, 0x09, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x43,
-	0x6f, 0x6d, 0x6d, 0x12, 0x2f, 0x0a, 0x02, 0x6d, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32,
-	0x1f, 0x2e, 0x74, 0x77, 0x74, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2e, 0x50, 0x72, 0x6f, 0x78, 0x79,
-	0x43, 0x6f, 0x6d, 0x6d, 0x2e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65,
-	0x52, 0x02, 0x6d, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x04, 0x52, 0x05, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x12, 0x1e, 0x0a, 0x0a, 0x63, 0x6f,
-	0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x03, 0x20, 0x01, 0x28, 0x04, 0x52, 0x0a,
-	0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x10, 0x0a, 0x03, 0x73, 0x65,
-	0x71, 0x18, 0x04, 0x20, 0x01, 0x28, 0x04, 0x52, 0x03, 0x73, 0x65, 0x71, 0x12, 0x18, 0x0a, 0x07,
-	0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x61,
-	0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x12, 0x12, 0x0a, 0x04, 0x70, 0x6f, 0x72, 0x74, 0x18, 0x06,
-	0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x70, 0x6f, 0x72, 0x74, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61,
-	0x74, 0x61, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0x66,
-	0x0a, 0x0b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x54, 0x79, 0x70, 0x65, 0x12, 0x0d, 0x0a,
-	0x09, 0x44, 0x41, 0x54, 0x41, 0x5f, 0x44, 0x4f, 0x57, 0x4e, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07,
-	0x44, 0x41, 0x54, 0x41, 0x5f, 0x55, 0x50, 0x10, 0x01, 0x12, 0x0d, 0x0a, 0x09, 0x4f, 0x50, 0x45,
-	0x4e, 0x5f, 0x43, 0x4f, 0x4e, 0x4e, 0x10, 0x02, 0x12, 0x10, 0x0a, 0x0c, 0x43, 0x4c, 0x4f, 0x53,
-	0x45, 0x5f, 0x43, 0x4f, 0x4e, 0x4e, 0x5f, 0x53, 0x10, 0x03, 0x12, 0x10, 0x0a, 0x0c, 0x43, 0x4c,
-	0x4f, 0x53, 0x45, 0x5f, 0x43, 0x4f, 0x4e, 0x4e, 0x5f, 0x43, 0x10, 0x04, 0x12, 0x08, 0x0a, 0x04,
-	0x50, 0x49, 0x4e, 0x47, 0x10, 0x05, 0x42, 0x14, 0x5a, 0x12, 0x6b, 0x65, 0x74, 0x72, 0x79, 0x2e,
-	0x6e, 0x65, 0x74, 0x2f, 0x74, 0x77, 0x74, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x33,
-}
+const file_twt_proto_rawDesc = "" +
+	"\n" +
+	"\ttwt.proto\x12\btwtproto\"\xc9\x02\n" +
+	"\tProxyComm\x12/\n" +
+	"\x02mt\x18\x01 \x01(\x0e2\x1f.twtproto.ProxyComm.MessageTypeR\x02mt\x12\x14\n" +
+	"\x05proxy\x18\x02 \x01(\x04R\x05proxy\x12\x1e\n" +
+	"\n" +
+	"connection\x18\x03 \x01(\x04R\n" +
+	"connection\x12\x10\n" +
+	"\x03seq\x18\x04 \x01(\x04R\x03seq\x12\x18\n" +
+	"\aaddress\x18\x05 \x01(\tR\aaddress\x12\x12\n" +
+	"\x04port\x18\x06 \x01(\rR\x04port\x12\x12\n" +
+	"\x04data\x18\a \x01(\fR\x04data\"\x80\x01\n" +
+	"\vMessageType\x12\r\n" +
+	"\tDATA_DOWN\x10\x00\x12\v\n" +
+	"\aDATA_UP\x10\x01\x12\f\n" +
+	"\bACK_DOWN\x10\x02\x12\n" +
+	"\n" +
+	"\x06ACK_UP\x10\x03\x12\r\n" +
+	"\tOPEN_CONN\x10\x04\x12\x10\n" +
+	"\fCLOSE_CONN_S\x10\x05\x12\x10\n" +
+	"\fCLOSE_CONN_C\x10\x06\x12\b\n" +
+	"\x04PING\x10\aB\x15Z\x13palecci.cz/twtprotob\x06proto3"
 
 var (
 	file_twt_proto_rawDescOnce sync.Once
-	file_twt_proto_rawDescData = file_twt_proto_rawDesc
+	file_twt_proto_rawDescData []byte
 )
 
 func file_twt_proto_rawDescGZIP() []byte {
 	file_twt_proto_rawDescOnce.Do(func() {
-		file_twt_proto_rawDescData = protoimpl.X.CompressGZIP(file_twt_proto_rawDescData)
+		file_twt_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_twt_proto_rawDesc), len(file_twt_proto_rawDesc)))
 	})
 	return file_twt_proto_rawDescData
 }
 
 var file_twt_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_twt_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_twt_proto_goTypes = []interface{}{
+var file_twt_proto_goTypes = []any{
 	(ProxyComm_MessageType)(0), // 0: twtproto.ProxyComm.MessageType
 	(*ProxyComm)(nil),          // 1: twtproto.ProxyComm
 }
@@ -233,25 +235,11 @@ func file_twt_proto_init() {
 	if File_twt_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_twt_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ProxyComm); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_twt_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_twt_proto_rawDesc), len(file_twt_proto_rawDesc)),
 			NumEnums:      1,
 			NumMessages:   1,
 			NumExtensions: 0,
@@ -263,7 +251,6 @@ func file_twt_proto_init() {
 		MessageInfos:      file_twt_proto_msgTypes,
 	}.Build()
 	File_twt_proto = out.File
-	file_twt_proto_rawDesc = nil
 	file_twt_proto_goTypes = nil
 	file_twt_proto_depIdxs = nil
 }
